src/fls_show/pyproject.toml
[project]
name = "fls-show"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Coding-AI-with-html", email = "krekeris440@gmail.com" }
]
requires-python = ">=3.12"
dependencies = []

[project.gui-scripts]
fls-show = "fls_show.fls_show:main"



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pydantic>=2.11.1",
    "pystray>=0.19.5",
]
